+++ R/utils.R
@@ -109,13 +109,15 @@ coerceValue = function(val, old) {
 #' @param data The original data object used in the DataTable.
 #' @param info The information about the edited cells. It should be obtained
 #'   from \code{input$tableId_cell_edit} from Shiny, and is a data frame
-#'   containing columns \code{row}, \code{column}, and \code{value}.
+#'   containing columns \code{row}, \code{col}, and \code{value}.
 #' @param rownames Whether row names are displayed in the table.
 #' @param proxy,resetPaging,... (Optional) If \code{proxy} is provided, it must
 #'   be either a character string of the output ID of the table or a proxy
 #'   object created from \code{\link{dataTableProxy}()}, and the rest of
 #'   arguments are passed to \code{\link{replaceData}()} to update the data in a
 #'   DataTable instance in a Shiny app.
+#' @note For factor columns, new levels would be automatically added when necessary
+#'   to avoid \code{NA} coercing.
 #' @return The updated data object.
 #' @export
 editData = function(data, info, proxy = NULL, rownames = TRUE, resetPaging = FALSE, ...) {
@@ -126,6 +128,10 @@ editData = function(data, info, proxy = NULL, rownames = TRUE, resetPaging = FAL
     if (j == 0) {
       rownames(data)[i] = v
     } else {
+      # allow add new factor levels
+      if (is.factor(data[[j]]) && !all(v %in% levels(data[[j]]))) {
+        levels(data[[j]]) <- unique(c(levels(data[[j]]), v))
+      }
       data[i, j] = coerceValue(v, data[i, j, drop = TRUE])
     }
   }