+++ tests/testit/test-datatables.R
@@ -173,3 +173,53 @@ local({
     (names(out$x$selection) %==% c('mode', 'selected', 'target', 'selectable'))
   })
 })
+
+assert('DT2BSClass() keeps user-defined classes', {
+  (DT:::DT2BSClass(c('table-condensed stripe', 'foo')) %==% 'table table-striped table-condensed foo')
+})
+
+assert('clear message when options$search is illegal', {
+  out = try(datatable(data = iris, options = list(search = TRUE)), silent = TRUE)
+  (inherits(out, 'try-error'))
+  (grepl('must be NULL or a list', out[1L], fixed = TRUE))
+})
+
+assert('warn autoHideNavigation if no pageLength', {
+  (has_warning(
+    datatable(head(iris, 5), autoHideNavigation = TRUE)
+  ))
+  (!has_warning(
+    datatable(head(iris, 5), autoHideNavigation = TRUE, options = list(pageLength = 20))
+  ))
+})
+
+assert("colDefsTgtHandle() works", {
+  cols = c("A", "B", "C")
+  (colDefsTgtHandle(NULL, cols) %==% list())
+  (colDefsTgtHandle(list(), cols) %==% list())
+  (has_error(colDefsTgtHandle("abc", cols)))
+  (has_error(colDefsTgtHandle(list("abc"), cols)))
+  defs = list(
+    list(1, targets = "_all"),
+    list(2, targets = 1L),
+    list(3, targets = "B"),
+    list(4, targets = c("A", "_all")),
+    list(5, targets = list(c("A", "C"), "_all")),
+    list(6, targets = list(1L, "_all")),
+    list(7, targets = list(1L, "C")),
+    list(8, targets = list(1L, "B", "_all")),
+    list(9, targets = list(1L, c("_all", "C")))
+  )
+  res = list(
+    list(1, targets = "_all"),
+    list(2, targets = 1L),
+    list(3, targets = 1L),
+    list(4, targets = "_all"),
+    list(5, targets = list(c(0L, 2L), "_all")),
+    list(6, targets = list(1L, "_all")),
+    list(7, targets = list(1L, 2L)),
+    list(8, targets = list(1L, 1L, "_all")),
+    list(9, targets = list(1L, "_all"))
+  )
+  (colDefsTgtHandle(defs, cols) %==% res)
+})